refactor: move floating buttons to client component
